+++ DTO/DtoListaPrecioMatPrProv.cs
@@ -6,7 +6,7 @@
         public int IdProveedor { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Telefono { get; set; }
+        public decimal Telefono { get; set; }
         public bool? Activo { get; set; }
         public string Localidad { get; set; }
         public string Provincia { get; set; } = null!;
